Refactor faucet_event_client into python_lib
